+++ controllers/account.go
@@ -16,6 +16,7 @@ import (
 	"github.com/twreporter/go-api/models"
 	"github.com/twreporter/go-api/storage"
 	"github.com/twreporter/go-api/utils"
+	"github.com/twreporter/go-api/configs/constants"
 )
 
 const idTokenExpiration = 60 * 60 * 24 * 30 * 6
@@ -23,10 +24,11 @@ const idTokenExpiration = 60 * 60 * 24 * 30 * 6
 var defaultRedirectPage = "https://www.twreporter.org/"
 var defaultPath = "/"
 
-func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath string) (int, gin.H, error) {
+func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath string, isCheckActivate bool) (int, gin.H, error) {
 	// SignInBody is to store POST body
 	type SignInBody struct {
 		Email            string `json:"email" form:"email" binding:"required"`
+		OnBoarding       string `json:"onboarding" form:"onboarding"`
 		Destination      string `json:"destination" form:"destination"`
 		ErrorRedirection string `json:"errorRedirection" form:"errorRedirection"`
 	}
@@ -39,6 +41,7 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 	var ra models.ReporterAccount
 	var statusCode int
 	var signIn SignInBody
+	var destination string
 
 	switch globals.Conf.Environment {
 	case "development":
@@ -55,6 +58,7 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 	if err = c.Bind(&signIn); err != nil {
 		return http.StatusBadRequest, gin.H{"status": "fail", "data": SignInBody{
 			Email:            "email is required",
+			OnBoarding:       "onboaring url is optional",
 			Destination:      "destination is optional",
 			ErrorRedirection: "errorRedirection is optional",
 		}}, nil
@@ -106,7 +110,7 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 		// the user record is not existed
 		if err != nil {
 			// create records both in reporter_accounts and users table
-			_, err = mc.Storage.InsertUserByReporterAccount(ra)
+			_, err := mc.Storage.InsertUserByReporterAccount(ra)
 			if err != nil {
 				return http.StatusInternalServerError, gin.H{"status": "error", "message": "Inserting new record into DB occurs error"}, nil
 			}
@@ -120,8 +124,22 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 				return http.StatusInternalServerError, gin.H{"status": "error", "message": "Inserting new record into DB occurs error"}, nil
 			}
 		}
-
 		statusCode = http.StatusCreated
+	}
+
+	// redirect to onboarding page if user is not activated
+	destination = signIn.Destination
+	matchedUser, err = mc.Storage.GetUserByEmail(email)
+	if err != nil {
+		fmt.Printf("cannot get user by email %s, use destination directly", email)
+	} else {
+		isActivate := matchedUser.Activated.Valid
+		if isCheckActivate && !isActivate {
+			destination = fmt.Sprintf("%s?destination=%s",
+				signIn.OnBoarding,
+				url.QueryEscape(signIn.Destination),
+			)
+		}
 	}
 
 	// send activation email
@@ -133,7 +151,7 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 			globals.Conf.App.Port,
 			url.QueryEscape(email),
 			url.QueryEscape(activeToken),
-			url.QueryEscape(signIn.Destination),
+			url.QueryEscape(destination),
 			url.QueryEscape(signIn.ErrorRedirection),
 		),
 	}, fmt.Sprintf("http://localhost:%s/v1/%s", globals.LocalhostPort, sendMailRoutePath))
@@ -144,6 +162,7 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 
 	return statusCode, gin.H{"status": "success", "data": SignInBody{
 		Email:            email,
+		OnBoarding:       signIn.OnBoarding,
 		Destination:      signIn.Destination,
 		ErrorRedirection: signIn.ErrorRedirection,
 	}}, nil
@@ -151,12 +170,12 @@ func (mc *MembershipController) AuthByEmail(c *gin.Context, sendMailRoutePath st
 
 // SignInV2 - send email containing sign-in information to the client
 func (mc *MembershipController) SignInV2(c *gin.Context) (int, gin.H, error) {
-	return mc.AuthByEmail(c, globals.SendActivationRoutePath)
+	return mc.AuthByEmail(c, globals.SendActivationRoutePath, true)
 }
 
 // AuthenticateV2 - send email containing authenticate information to the client
 func (mc *MembershipController) AuthenticateV2(c *gin.Context) (int, gin.H, error) {
-	return mc.AuthByEmail(c, globals.SendAuthenticationRoutePath)
+	return mc.AuthByEmail(c, globals.SendAuthenticationRoutePath, false)
 }
 
 // ActivateV2 - validate the reporter account
@@ -245,6 +264,12 @@ func (mc *MembershipController) ActivateV2(c *gin.Context) {
 	u.RawQuery = parameters.Encode()
 	destination = u.String()
 
+	var activatedString string
+	if user.Activated.Valid && !user.Activated.Time.IsZero() {
+		activatedString = user.Activated.Time.Format(time.RFC3339)
+	}
+
+	c.SetCookie("activated", activatedString, idTokenExpiration, defaultPath, defaultDomain, secure, true)
 	c.SetCookie("id_token", idToken, idTokenExpiration, defaultPath, defaultDomain, secure, true)
 	c.Redirect(http.StatusTemporaryRedirect, destination)
 }
@@ -305,4 +330,55 @@ func (mc *MembershipController) TokenInvalidate(c *gin.Context) {
 
 	c.SetCookie(cookieName, "", invalidateExp, defaultPath, defaultDomain, u.Scheme == "https", true)
 	c.Redirect(http.StatusTemporaryRedirect, destination)
+}
+
+// Onboarding set user preference & activated
+func (mc *MembershipController) Onboarding(c *gin.Context) (int, gin.H, error) {
+	destination := c.Query("destination")
+	u, _ := url.Parse(destination)
+	userID := c.Param("userID")
+
+	var preferences models.UserPreference
+	err := c.BindJSON(&preferences)
+	if err != nil {
+		fmt.Println("Error decoding JSON:", err)
+		return toResponse(err)
+	}
+
+	// Convert maillist values using the mapping array
+	maillists := make([]string, 0)
+	for _, maillist := range preferences.Maillist {
+		convertedMaillist, exists := globals.Conf.Mailchimp.InterestIDs[maillist]
+		if !exists {
+			return http.StatusBadRequest, gin.H{"status": "error", "message": "invalid maillist value"}, errors.New("Invalid maillist value")
+		}
+		maillists = append(maillists, convertedMaillist)
+	}
+
+	// Call UpdateReadPreferenceOfUser to save the preferences.ReadPreference to DB
+	if err = mc.Storage.UpdateReadPreferenceOfUser(userID, preferences.ReadPreference); err != nil {
+		return toResponse(err)
+	}
+
+	// Call CreateMaillistOfUser to save the preferences.Maillist to DB
+	if err = mc.Storage.CreateMaillistOfUser(userID, maillists); err != nil {
+		return toResponse(err)
+	}
+
+	user, _ := mc.Storage.GetUserByID(fmt.Sprint(userID))
+	go mc.sendAssignRoleMail(constants.RoleExplorer, user.Email.String)
+
+	if user.Activated.Valid && !user.Activated.Time.IsZero() {
+		activatedString := user.Activated.Time.Format(time.RFC3339)
+		activatedExpiration := idTokenExpiration
+		defaultDomain := globals.Conf.App.Domain
+		// Determine cookie property
+		secure := false
+		if "https" == u.Scheme {
+			secure = true
+		}
+
+		c.SetCookie("activated", activatedString, activatedExpiration, defaultPath, defaultDomain, secure, true)
+	}
+	return http.StatusCreated, gin.H{"status": "ok", "record": preferences}, nil
 }