+++ controllers/mail.go
@@ -14,6 +14,7 @@ import (
 	"github.com/pkg/errors"
 	"gopkg.in/guregu/null.v3"
 
+	"github.com/twreporter/go-api/globals"
 	"github.com/twreporter/go-api/services"
 	"github.com/twreporter/go-api/utils"
 )
@@ -34,6 +35,11 @@ type donationSuccessReqBody struct {
 	IsAutoPay         bool     `json:"is_auto_pay"`
 	Name              string   `json:"name"`
 	OrderNumber       string   `json:"order_number" binding:"required"`
+}
+
+type assignRoleReqBody struct {
+	RoleKey string `json:"role" binding:"required"`
+	Email   string `json:"email" binding:"required"`
 }
 
 // NewMailController is used to new *MailController
@@ -166,6 +172,63 @@ func (contrl *MailController) SendDonationSuccessMail(c *gin.Context) (int, gin.
 	}
 
 	return http.StatusNoContent, gin.H{}, nil
+}
+
+func (contrl *MailController) sendRoleMail(c *gin.Context, subject, templateName string) (int, gin.H, error) {
+	var err error
+	var mailBody string
+	var out bytes.Buffer
+	var reqBody assignRoleReqBody
+
+	if failData, err := bindRequestJSONBody(c, &reqBody); err != nil {
+		return http.StatusBadRequest, gin.H{"status": "fail", "data": failData}, nil
+	}
+
+	if globals.Conf.Features.EnableRolemail {
+		// If Features.EnableRolemail is true, send email through mail service
+		var templateData = struct {
+			Subject     string
+			CurrentYear string
+		}{
+			subject,
+			fmt.Sprintf("%d", time.Now().Year()),
+		}
+
+		if err = contrl.HTMLTemplate.ExecuteTemplate(&out, templateName, templateData); err != nil {
+			return http.StatusInternalServerError, gin.H{"status": "error", "message": "can not create assign role mail body"}, errors.WithStack(err)
+		}
+
+		mailBody = out.String()
+
+		if err = contrl.MailService.Send(reqBody.Email, subject, mailBody); err != nil {
+			return http.StatusInternalServerError, gin.H{"status": "error", "message": fmt.Sprintf("can not send role mail to %s", reqBody.Email)}, err
+		}
+	} else {
+		// If Features.EnableRolemail is false, print a log only
+		fmt.Printf("Mail not sent due to feature toggle (Features.EnableRolemail) is (%v): [%s] %s\n", globals.Conf.Features.EnableRolemail, reqBody.Email, subject)
+	}
+
+	return http.StatusNoContent, gin.H{}, nil
+}
+
+func (contrl *MailController) SendRoleExplorerMail(c *gin.Context) (int, gin.H, error) {
+	const subject = "歡迎您成為探索者，與《報導者》一起看見世界上正在發生的重要的事"
+	return contrl.sendRoleMail(c, subject, "role-explorer.tmpl")
+}
+
+func (contrl *MailController) SendRoleActiontakerMail(c *gin.Context) (int, gin.H, error) {
+	const subject = "歡迎成為「行動者」，這些是我們為你提供的服務"
+	return contrl.sendRoleMail(c, subject, "role-actiontaker.tmpl")
+}
+
+func (contrl *MailController) SendRoleTrailblazerMail(c *gin.Context) (int, gin.H, error) {
+	const subject = "歡迎成為「開創者」，這些是我們為你提供的服務"
+	return contrl.sendRoleMail(c, subject, "role-trailblazer.tmpl")
+}
+
+func (contrl *MailController) SendRoleDowngradeMail(c *gin.Context) (int, gin.H, error) {
+	const subject = "方案身分異動通知"
+	return contrl.sendRoleMail(c, subject, "role-downgrade.tmpl")
 }
 
 func postMailServiceEndpoint(reqBody interface{}, endpoint string) error {