+++ routers/router.go
@@ -138,6 +138,10 @@ func SetupRouter(cf *controllers.ControllerFactory) (engine *gin.Engine) {
 	v1Group.POST(fmt.Sprintf("/%s", globals.SendActivationRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendActivation))
 	v1Group.POST(fmt.Sprintf("/%s", globals.SendAuthenticationRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendAuthentication))
 	v1Group.POST(fmt.Sprintf("/%s", globals.SendSuccessDonationRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendDonationSuccessMail))
+	v1Group.POST(fmt.Sprintf("/%s", globals.SendRoleExplorerRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendRoleExplorerMail))
+	v1Group.POST(fmt.Sprintf("/%s", globals.SendRoleActiontakerRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendRoleActiontakerMail))
+	v1Group.POST(fmt.Sprintf("/%s", globals.SendRoleTrailblazerRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendRoleTrailblazerMail))
+	v1Group.POST(fmt.Sprintf("/%s", globals.SendRoleDowngradeRoutePath), mailMiddleware.ValidateAuthorization(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mailContrl.SendRoleDowngradeMail))
 
 	v2Group := engine.Group("/v2")
 	ncV2 := cf.GetNewsV2Controller()
@@ -160,6 +164,9 @@ func SetupRouter(cf *controllers.ControllerFactory) (engine *gin.Engine) {
 		}
 		c.AbortWithStatus(http.StatusNotFound)
 	})
+
+	v2Group.POST("/users/:userID", middlewares.ValidateAuthorization(), middlewares.ValidateUserID(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mc.SetUser))
+	v2Group.GET("/users/:userID", middlewares.ValidateAuthorization(), middlewares.ValidateUserID(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mc.GetUser))
 
 	// =============================
 	// v2 oauth endpoints
@@ -192,5 +199,6 @@ func SetupRouter(cf *controllers.ControllerFactory) (engine *gin.Engine) {
 	v2AuthGroup.GET("/activate", middlewares.SetCacheControl("no-store"), mc.ActivateV2)
 	v2AuthGroup.POST("/token", middlewares.ValidateAuthentication(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mc.TokenDispatch))
 	v2AuthGroup.GET("/logout", mc.TokenInvalidate)
+	v2Group.POST("/onboarding/:userID", middlewares.ValidateAuthorization(), middlewares.ValidateUserID(), middlewares.SetCacheControl("no-store"), ginResponseWrapper(mc.Onboarding))
 	return
 }